misc: code refactor

+++ runtime/thread/local_vars.go
@@ -6,10 +6,10 @@ func newLocalVars(maxLocals uint) LocalVars {
 	return make([]Slot, maxLocals)
 }
 
-func (l LocalVars) SetInt(index int, val int32) {
+func (l LocalVars) SetInt(index uint8, val int32) {
 	l[index].num = val
 }
 
-func (l LocalVars) GetInt(index int) int32 {
+func (l LocalVars) GetInt(index uint8) int32 {
 	return l[index].num
 }
